+++ packages/core/src/types/index.ts
@@ -3,5 +3,4 @@ export * from "./response.types";
 export * from "./request.types";
 export * from "./schema.types";
 export * from "./client.types";
-export * from "./middleware.types";
 export * from "./instance.types";